% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolR.R
\name{logger}
\alias{logger}
\title{Logger}
\usage{
logger(site, wtd_path, aws_path, references)
}
\arguments{
\item{site}{Site Name}

\item{wtd_path}{Path to .xle Solisnt water table data folder}

\item{aws_path}{Path to automatic weather station folder}

\item{references}{Path to .csv of water table depth metadata}
}
\value{
Plots and csv data saved to SolR folder within the water table data folder
}
\description{
Extract, transform and graph Solinst water table data
}
